게시판 4개 검색 수정, 글쓰기, 조회수, 페이징 완료

+++ Project/src/main/java/com/smhrd/news/NewsVO.java
@@ -10,11 +10,11 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class NewsVO {
 
-    private Long news_idx;            // 공지 식별자
+    private int news_idx;            // 공지 식별자
     private String news_title;        // 제목
     private String news_content;      // 내용 (CLOB)
     private String news_file;         // 첨부 파일명
-    private Long news_views;          // 조회수
+    private int news_views;          // 조회수
     private Timestamp created_at;     // 생성일
     private Timestamp updated_at;     // 수정일
     private String admin_id;          // 작성자 아이디 (FK)
